Add: optionally filter ActivityPub hashtags
It is conventional on the ActivityPub fediverse to be liberal with the
use of hashtags. These are typically listed at the end of messages so
that they can be surfaced appropriately on clients.

This is at odds with other social platforms, where hashtags are either
against social norms (Twitter) or they aren't supported (Bluesky).

This adds a new `removeTrailingHashtags` configuration for the
ActivityPub trigger to filter trailing hashtags out.

Because hashtags in the body of messages are often part of the semantics
of the text (i.e. "what's up #ttrpg players") we are leaving these
untouched to avoid distorting the meaning of the source messages.

For more information:

https://github.com/actionslaw/actionslaw-action/pull/2